requirements.txt
amqp==5.1.1
asgiref==3.5.0
Django==3.2.7
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
Pillow==9.1.0
python-dotenv==0.20.0
django-jazzmin==2.4.8
wrapt==1.14.0
psycopg2-binary==2.9.3
celery==5.2.6
redis==4.2.2
python-steam-api==1.0.8
beautifulsoup4